Type conversion fixed
В данном случае, по условию задачи, переменные i и j имеют тип integer.
В то же время, выражение i*j / i+2 может не быть целым числом, соответственно, иметь тип отличный от integer.
При вычислении целых чисел в итоге получается целое число (подразумевается тип).
Соответственно, есть два решения проблемы. Первый и менее эффективный - представить внутри выражения  часть операнда деления в роли дробного (вещественного) числа.
Например, представить 2 как 2.0. Но в этом случае мы получим дополнительное действие с дробным числом.
Второй вариант, который был использован, - явное приведение типа числителя или знаменателя к дробному.

+++ src/ip94/dolgova/Lab1.java
@@ -24,7 +24,7 @@ public class Lab1 {
 		    System.out.println("Error. Zero division is not allowed");
 		    break label;
 		} else {
-		    s += (i * j / (i + C));
+		    s += ((float) i * j / (i + C));
 		}
 	    }
